requirements.txt
boto3
python-dotenv
psycopg2-binary 
pandas 
sqlalchemy
